requirements.txt
torch
openai-whisper
transformers
datasets
arabert
fastapi
uvicorn
librosa
soundfile
scipy
numpy
webrtcvad
sounddevice
